requirements/base.txt
Django==1.9.1
Pillow==3.1.0.rc1
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.19
coverage==4.1b1
distribute==0.6.24
django-celery==3.1.17
kombu==3.0.33
psycopg2==2.6.1
pytz==2015.7
redis==2.10.5
